Add utility func for Error page to get title and subtitle
