+++ context/app/static/js/pages/Error/Error.jsx
@@ -4,23 +4,12 @@ import Typography from '@material-ui/core/Typography';
 
 import ErrorBody from 'js/components/error/ErrorBody';
 import { Background, StyledPaper, StyledTypography } from './style';
-
-const errorTitle = {
-  400: 'Bad Request',
-  401: 'Unauthorized',
-  403: 'Access Denied',
-  404: 'Page Not Found',
-  504: 'Gateway Timeout',
-  500: 'Internal Server Error',
-};
+import { getErrorTitleAndSubtitle } from './utils';
 
 function Error(props) {
   const { errorCode, isGlobus401, isMaintenancePage } = props;
 
-  const title = isMaintenancePage ? 'Portal Maintenance' : errorTitle[errorCode];
-  const subtitle = isMaintenancePage
-    ? 'Portal unavailable for scheduled maintenance.'
-    : `HTTP Error ${errorCode}: ${title}`;
+  const { title, subtitle } = getErrorTitleAndSubtitle(errorCode, isMaintenancePage);
 
   return (
     <Background>
